// src/app/features/user/interfaces/user.interface.ts
export interface User {
    height: number;
    id: string;
    description: string;
    imgUrl: string;
    name: string;
    type: string;
    weight: number;
    contact:number;
   
  }
  export interface Items {
    height: number;
    id: string;
    description: string;
    imgUrl: string;
    name: string;
    type: string;
    weight: number;
    contact:number;
   
  }

  export interface UserComponent {}
  export interface ItemsComponent {}

  